fix SQL Server false configuration error
When `TrustServerCertificate=yes` is used in the connection string it produces the following error because `TrustServerCertificate` contains the word `Server`:

```
datadog_checks.base.errors.ConfigurationError: SERVER has been provided both in the connection string and as a configuration option (host), please specify it only once
```

Add a new parsing step that explicitly parses out the connection string and provides a map where the keys are all of the configured settings. This is more robust than doing a basic substring check.